Move non-tests out of the stormpot test package
The `stormpot` test package now only contain tests that rely on implementation details.
