+++ src/test/java/blackbox/AllocatorBasedPoolTest.java
@@ -19,13 +19,13 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
-import stormpot.AlloKit;
+import testkits.AlloKit;
 import stormpot.Allocator;
 import stormpot.Completion;
 import stormpot.Expiration;
-import stormpot.FixedMeanMetricsRecorder;
-import stormpot.GenericPoolable;
-import stormpot.LastSampleMetricsRecorder;
+import testkits.FixedMeanMetricsRecorder;
+import testkits.GenericPoolable;
+import testkits.LastSampleMetricsRecorder;
 import stormpot.ManagedPool;
 import stormpot.Pool;
 import stormpot.PoolBuilder;
@@ -34,8 +34,8 @@ import stormpot.PoolTap;
 import stormpot.Poolable;
 import stormpot.Reallocator;
 import stormpot.Slot;
-import stormpot.SomeRandomException;
-import stormpot.SomeRandomThrowable;
+import testkits.SomeRandomException;
+import testkits.SomeRandomThrowable;
 import stormpot.Timeout;
 
 import java.lang.ref.WeakReference;
@@ -60,39 +60,39 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
-import static stormpot.AlloKit.$countDown;
-import static stormpot.AlloKit.$incrementAnd;
-import static stormpot.AlloKit.$new;
-import static stormpot.AlloKit.$null;
-import static stormpot.AlloKit.$observeNull;
-import static stormpot.AlloKit.$sleep;
-import static stormpot.AlloKit.$sync;
-import static stormpot.AlloKit.$throw;
-import static stormpot.AlloKit.CountingAllocator;
-import static stormpot.AlloKit.alloc;
-import static stormpot.AlloKit.allocator;
-import static stormpot.AlloKit.dealloc;
-import static stormpot.AlloKit.realloc;
-import static stormpot.AlloKit.reallocator;
-import static stormpot.ExpireKit.$age;
-import static stormpot.ExpireKit.$capture;
-import static stormpot.ExpireKit.$claimCount;
-import static stormpot.ExpireKit.$expired;
-import static stormpot.ExpireKit.$expiredIf;
-import static stormpot.ExpireKit.$explicitExpire;
-import static stormpot.ExpireKit.$fresh;
-import static stormpot.ExpireKit.$if;
-import static stormpot.ExpireKit.$poolable;
-import static stormpot.ExpireKit.$throwExpire;
-import static stormpot.ExpireKit.CountingExpiration;
-import static stormpot.ExpireKit.expire;
-import static stormpot.UnitKit.$delayedReleases;
-import static stormpot.UnitKit.claimRelease;
-import static stormpot.UnitKit.fork;
-import static stormpot.UnitKit.forkFuture;
-import static stormpot.UnitKit.join;
-import static stormpot.UnitKit.spinwait;
-import static stormpot.UnitKit.waitForThreadState;
+import static testkits.AlloKit.$countDown;
+import static testkits.AlloKit.$incrementAnd;
+import static testkits.AlloKit.$new;
+import static testkits.AlloKit.$null;
+import static testkits.AlloKit.$observeNull;
+import static testkits.AlloKit.$sleep;
+import static testkits.AlloKit.$sync;
+import static testkits.AlloKit.$throw;
+import static testkits.AlloKit.CountingAllocator;
+import static testkits.AlloKit.alloc;
+import static testkits.AlloKit.allocator;
+import static testkits.AlloKit.dealloc;
+import static testkits.AlloKit.realloc;
+import static testkits.AlloKit.reallocator;
+import static testkits.ExpireKit.$age;
+import static testkits.ExpireKit.$capture;
+import static testkits.ExpireKit.$claimCount;
+import static testkits.ExpireKit.$expired;
+import static testkits.ExpireKit.$expiredIf;
+import static testkits.ExpireKit.$explicitExpire;
+import static testkits.ExpireKit.$fresh;
+import static testkits.ExpireKit.$if;
+import static testkits.ExpireKit.$poolable;
+import static testkits.ExpireKit.$throwExpire;
+import static testkits.ExpireKit.CountingExpiration;
+import static testkits.ExpireKit.expire;
+import static testkits.UnitKit.$delayedReleases;
+import static testkits.UnitKit.claimRelease;
+import static testkits.UnitKit.fork;
+import static testkits.UnitKit.forkFuture;
+import static testkits.UnitKit.join;
+import static testkits.UnitKit.spinwait;
+import static testkits.UnitKit.waitForThreadState;
 
 /**
  * This is the generic test for Pool implementations. The test ensures that
