+++ src/test/java/blackbox/ThreadBasedPoolTest.java
@@ -18,9 +18,9 @@ package blackbox;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
-import stormpot.ExpireKit;
-import stormpot.GenericPoolable;
-import stormpot.LastSampleMetricsRecorder;
+import testkits.ExpireKit;
+import testkits.GenericPoolable;
+import testkits.LastSampleMetricsRecorder;
 import stormpot.ManagedPool;
 import stormpot.Pool;
 import stormpot.PoolBuilder;
@@ -46,30 +46,30 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
-import static stormpot.AlloKit.$acquire;
-import static stormpot.AlloKit.$countDown;
-import static stormpot.AlloKit.$if;
-import static stormpot.AlloKit.$new;
-import static stormpot.AlloKit.$null;
-import static stormpot.AlloKit.$throw;
-import static stormpot.AlloKit.Action;
-import static stormpot.AlloKit.CountingReallocator;
-import static stormpot.AlloKit.alloc;
-import static stormpot.AlloKit.allocator;
-import static stormpot.AlloKit.realloc;
-import static stormpot.AlloKit.reallocator;
-import static stormpot.ExpireKit.$countDown;
-import static stormpot.ExpireKit.$expired;
-import static stormpot.ExpireKit.$expiredIf;
-import static stormpot.ExpireKit.$fresh;
-import static stormpot.ExpireKit.$throwExpire;
-import static stormpot.ExpireKit.CountingExpiration;
-import static stormpot.ExpireKit.expire;
-import static stormpot.UnitKit.$claim;
-import static stormpot.UnitKit.$claimRelease;
-import static stormpot.UnitKit.capture;
-import static stormpot.UnitKit.forkFuture;
-import static stormpot.UnitKit.spinwait;
+import static testkits.AlloKit.$acquire;
+import static testkits.AlloKit.$countDown;
+import static testkits.AlloKit.$if;
+import static testkits.AlloKit.$new;
+import static testkits.AlloKit.$null;
+import static testkits.AlloKit.$throw;
+import static testkits.AlloKit.Action;
+import static testkits.AlloKit.CountingReallocator;
+import static testkits.AlloKit.alloc;
+import static testkits.AlloKit.allocator;
+import static testkits.AlloKit.realloc;
+import static testkits.AlloKit.reallocator;
+import static testkits.ExpireKit.$countDown;
+import static testkits.ExpireKit.$expired;
+import static testkits.ExpireKit.$expiredIf;
+import static testkits.ExpireKit.$fresh;
+import static testkits.ExpireKit.$throwExpire;
+import static testkits.ExpireKit.CountingExpiration;
+import static testkits.ExpireKit.expire;
+import static testkits.UnitKit.$claim;
+import static testkits.UnitKit.$claimRelease;
+import static testkits.UnitKit.capture;
+import static testkits.UnitKit.forkFuture;
+import static testkits.UnitKit.spinwait;
 
 abstract class ThreadBasedPoolTest extends AllocatorBasedPoolTest {
