+++ src/test/java/examples/ThreadLocalWeakRefTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package stormpot;
+package examples;
 
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
