+++ src/test/java/stormpot/RefillPileTest.java
@@ -19,6 +19,7 @@ import extensions.ExecutorExtension;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
+import testkits.GenericPoolable;
 
 import java.util.HashSet;
 import java.util.Set;
