+++ src/test/java/testkits/FixedMeanMetricsRecorder.java
@@ -13,7 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package stormpot;
+package testkits;
+
+import stormpot.MetricsRecorder;
 
 import static java.lang.Double.NaN;
